backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
asyncpg==0.31.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.12.5
python-multipart==0.0.22
python-dotenv==1.2.2
